add ids for electrodes and their recipes
